+++ OpenNeighborhood/src/Panels/ContentsPanel.cpp
@@ -46,7 +46,7 @@ void ContentsPanel::OnRender()
 	;
 
 	ImGui::SetNextWindowPos(ImVec2(m_Margin, m_Margin * 2.0f + pathPanelHeight));
-	ImGui::SetNextWindowSize(ImVec2((float)m_WindowWidth - m_Margin * 2.0f, (float)m_WindowHeight - (m_Margin * 3.0f + pathPanelHeight)));
+	ImGui::SetNextWindowSize(ImVec2(m_WindowWidth - m_Margin * 2.0f, m_WindowHeight - (m_Margin * 3.0f + pathPanelHeight)));
 
 	ImGui::Begin("Contents Window", nullptr, windowFlags);
